requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
chromadb
openai
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
python-multipart
